add get_extents for Text shapes
this simply calls cairo_context.font_extents() and cairo_context.text_extents(...).
it returns the results inside a dict. the keys that correspond to the values returned
by .font_extents are:

    ascent
    descent
    font_height
    max_x_advance
    max_y_advance

and the ones returned by .text_extents are:

    x_bearing
    y_bearing
    text_width
    text_height
    x_advance
    y_advance

see the pycairo documentation for more details on what these values stand for.

http://cairographics.org/documentation/pycairo/3/reference/context.html#cairo.Context.font_extents
http://cairographics.org/documentation/pycairo/3/reference/context.html#cairo.Context.text_extents