add emoji support for Images and Texts
you have to supply your own emoji set by setting the
emoji_path attribute on a RenderList to the folder
that contains the pngs you want to use.

right now there's no error checking, so if an emoji
is missing or something else of that nature happens,
things will just crash and burn.