add gif duration detection
in Images, this is available as img.duration.

if you want to get the individual time for each
frame (in milliseconds), this is available as a list
in img.frame_durations.

this is also available as an utility
function, called get_gif_duration.

the utility function uses PIL/Pillow, as
that's apparently faster than imageio/FreeImage
when doing this.